+++ CHANGES.rst
@@ -9,6 +9,11 @@ ramp_fitting
 
 - Adding special case handler for GLS to handle one group ramps. [#97]
 
+- Updating how one group suppression and ZEROFRAME processing works with
+  multiprocessing, as well as fixing the multiprocessing failure. [#99]
+
+- Changing how ramp fitting handles fully saturated ramps. [#102]
+
 saturation
 ~~~~~~~~~~